backend/requirements.txt
fastapi[all]>=0.115.0
uvicorn[standard]>=0.34.0
gunicorn>=23.0.0
beanie>=1.29.0
motor>=3.6.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
authlib>=1.4.0
httpx>=0.28.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.18
redis[hiredis]>=5.2.0
pycrdt>=0.12.0
pycrdt-websocket>=0.16.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
pytest-xdist>=3.5.0
pytest-cov>=6.0.0
mongomock-motor>=0.0.34
python3-saml>=1.16.0
ruff>=0.15.0
